# openapi/mygpt_openapi_minimal.yaml
openapi: 3.1.0
info:
  title: Biwenger Mundial 2026 Actions API
  description: API privada de administración del optimizador Biwenger Mundial 2026.
    Requiere la clave API de administrador en X-API-Key y un user_id estable en cada
    ruta. user_id separa configuración, ratings y plantilla local, pero no autentica
    por sí solo. Las acciones de plantilla son simulaciones locales y requieren el
    confirmation_token devuelto por la API. No existe acceso a credenciales ni a la
    cuenta real de Biwenger.
  version: 0.2.0
paths:
  /api/users/{user_id}/status:
    get:
      summary: Status Route
      operationId: getStatus
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/rules:
    get:
      summary: Rules
      description: Return configurable example rules for this private league, not
        official Biwenger rules.
      operationId: getRules
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/settings:
    get:
      summary: Get Settings
      description: Retrieve the current active budget and starting lineup formation
        settings.
      operationId: getSettings
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Update Settings
      description: Update settings (e.g. budget limit, starting lineup formation)
        dynamically.
      operationId: updateSettings
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/sync:
    post:
      summary: Sync Database
      description: Trigger synchronization between the local database and the Biwenger
        API.
      operationId: syncPublicData
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Sync Database Via Get
      description: Compatibility alias for clients that call sync via GET.
      operationId: syncPublicDataViaGet
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/players:
    get:
      summary: Get Players
      description: Fetch players with optional query filters, sorting, and pagination.
      operationId: listPlayers
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      - name: position
        in: query
        required: false
        schema:
          description: 'Filter by position: GK, DEF, MID, FWD'
          title: Position
          type: string
          enum:
          - GK
          - DEF
          - MID
          - FWD
        description: 'Filter by position: GK, DEF, MID, FWD'
        style: form
        explode: true
      - name: team
        in: query
        required: false
        schema:
          description: Filter by country/national team name
          title: Team
          type: string
        description: Filter by country/national team name
        style: form
        explode: true
      - name: status
        in: query
        required: false
        schema:
          description: 'Filter by availability status: ok, doubtful, injured'
          title: Status
          type: string
          enum:
          - ok
          - doubtful
          - injured
          - suspended
          - no_disponible
        description: 'Filter by availability status: ok, doubtful, injured'
        style: form
        explode: true
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          description: 'Sort by: fixed_price, points, market_value, goals, assists'
          default: fixed_price
          title: Sort By
          enum:
          - fixed_price
          - points
          - market_value
          - goals
          - assists
          - name
          - team
        description: 'Sort by: fixed_price, points, market_value, goals, assists'
        style: form
        explode: true
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          description: Filter to show only players whose national teams are still
            active in the tournament
          default: true
          title: Active Only
        description: Filter to show only players whose national teams are still active
          in the tournament
        style: form
        explode: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
        style: form
        explode: true
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
        style: form
        explode: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/player/{player_id}:
    get:
      summary: Get Player
      description: Retrieve detailed stats for a specific player by ID.
      operationId: getPlayer
      security:
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          maxLength: 64
        description: Stable tenant identifier; this is not authentication.
      - name: player_id
        in: path
        required: true
        schema:
          type: string
          title: Player Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse:
      properties:
        ok:
          type: boolean
          title: Ok
        data:
          title: Data
        warnings:
          items:
            type: string
          type: array
          title: Warnings
        needs_review:
          type: boolean
          title: Needs Review
          default: false
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
      type: object
      required:
      - ok
      title: ApiResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
servers:
- url: https://biwenger-mcp.cestmail.com